+++ src/main/kotlin/com/github/anddd7/domain/stock/StockCoroutineRepository.kt
@@ -4,5 +4,5 @@ import com.github.anddd7.domain.Repository
 import java.math.BigDecimal
 
 interface StockCoroutineRepository : Repository {
-  suspend fun getStock(productId: Long): BigDecimal
+  suspend fun getStock(productId: Int): BigDecimal
 }
